+++ packages/vector_graphics/lib/src/filters/filter_shaders.dart
@@ -95,6 +95,8 @@ class FilterShaders {
             names.add('displacement_map');
           case 'feImage':
             names.add('image');
+          case 'feTurbulence':
+            names.add('turbulence');
         }
       }
     }